task
Create a base class called shape to store two double type values that could be used to compute
the area of rectangular and/or triangle. Two separate classes are derived from the base class for
the same. The input for the computation and the output is to be displayed in the base class. Make
the function to display as a virtual function and redefine this function in the derived class to suit
their requirement.